requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
czifile>=2019.7.2
cellpose<=3.0.0
opencv-python>=4.5.0
PyYAML>=5.4.0
scikit-image>=0.18.0
torch>=1.9.0
